Added guards against writing to read-only files Provided a CreateFilesEndpoint example in the Sample service (overrides)
